Fix calculation of density bands when cell seg data is in microns (ticket #10).
When inForm writes cell data in microns, it uses the slide origin for the cell location. This has to be subtracted out to get image coordinates.

Note: when read_cell_seg_data converts to microns it does not add the offset so the coordinate origin is at the image top left.